fix: refactors the order history cubit to have cancelorder method
