+++ lib/features/coin/presentation/cubits/order_history_cubit.dart
@@ -2,6 +2,7 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:polkadex/common/orders/domain/entities/order_entity.dart';
+import 'package:polkadex/common/orders/domain/usecases/cancel_order_usecase.dart';
 import 'package:polkadex/common/orders/domain/usecases/get_orders_usecase.dart';
 
 part 'order_history_state.dart';
@@ -9,10 +10,14 @@ part 'order_history_state.dart';
 class OrderHistoryCubit extends Cubit<OrderHistoryState> {
   OrderHistoryCubit({
     required GetOrdersUseCase getOrdersUseCase,
+    required CancelOrderUseCase cancelOrderUseCase,
   })  : _getOrdersUseCase = getOrdersUseCase,
+        _cancelOrderUseCase = cancelOrderUseCase,
         super(OrderHistoryInitial());
 
   final GetOrdersUseCase _getOrdersUseCase;
+  final CancelOrderUseCase _cancelOrderUseCase;
+
   List<OrderEntity> _allOrders = [];
 
   Future<void> getOrders(
@@ -52,6 +57,7 @@ class OrderHistoryCubit extends Cubit<OrderHistoryState> {
 
         emit(OrderHistoryLoaded(
           orders: _allOrders,
+          orderIdsLoading: [],
         ));
       },
     );
@@ -61,7 +67,9 @@ class OrderHistoryCubit extends Cubit<OrderHistoryState> {
     required List<Enum> filters,
     DateTimeRange? dateFilter,
   }) async {
+    final previousState = state;
     List<OrderEntity> _ordersFiltered = [..._allOrders];
+    List<String> _ordersIdLoading;
 
     if (dateFilter != null) {
       _ordersFiltered.removeWhere((order) =>
@@ -74,6 +82,57 @@ class OrderHistoryCubit extends Cubit<OrderHistoryState> {
           .removeWhere((order) => !filters.contains(order.orderSide));
     }
 
-    emit(OrderHistoryLoaded(orders: _ordersFiltered));
+    if (previousState is OrderHistoryLoaded) {
+      _ordersIdLoading = previousState.orderIdsLoading;
+    } else {
+      _ordersIdLoading = [];
+    }
+
+    emit(
+      OrderHistoryLoaded(
+          orders: _ordersFiltered, orderIdsLoading: _ordersIdLoading),
+    );
+  }
+
+  Future<bool> cancelOrder(
+    OrderEntity order,
+    String address,
+    String signature,
+  ) async {
+    final firstPreviousState = state;
+
+    if (firstPreviousState is OrderHistoryLoaded &&
+        firstPreviousState.orders.contains(order)) {
+      emit(OrderHistoryLoaded(
+        orders: [...firstPreviousState.orders],
+        orderIdsLoading: [
+          ...firstPreviousState.orderIdsLoading,
+          order.orderId,
+        ],
+      ));
+
+      final result = await _cancelOrderUseCase(
+        nonce: 0,
+        address: address,
+        orderId: order.orderId,
+        signature: signature,
+      );
+
+      final secondPreviousState = state;
+
+      if (secondPreviousState is OrderHistoryLoaded) {
+        emit(OrderHistoryLoaded(
+          orders: result.isRight()
+              ? ([...secondPreviousState.orders]..remove(order))
+              : [...secondPreviousState.orders],
+          orderIdsLoading: [...secondPreviousState.orderIdsLoading]
+            ..remove(order.orderId),
+        ));
+      }
+
+      return result.isRight();
+    } else {
+      return false;
+    }
   }
 }
