+++ lib/features/landing/presentation/cubits/list_orders_cubit/list_orders_cubit.dart
@@ -1,20 +1,16 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:polkadex/common/orders/domain/entities/order_entity.dart';
-import 'package:polkadex/common/orders/domain/usecases/cancel_order_usecase.dart';
 import 'package:polkadex/common/orders/domain/usecases/get_open_orders.dart';
 
 part 'list_orders_state.dart';
 
 class ListOrdersCubit extends Cubit<ListOrdersState> {
   ListOrdersCubit({
-    required CancelOrderUseCase cancelOrderUseCase,
     required GetOpenOrdersUseCase getOpenOrdersUseCase,
-  })  : _cancelOrderUseCase = cancelOrderUseCase,
-        _getOpenOrdersUseCase = getOpenOrdersUseCase,
+  })  : _getOpenOrdersUseCase = getOpenOrdersUseCase,
         super(ListOrdersInitial());
 
-  final CancelOrderUseCase _cancelOrderUseCase;
   final GetOpenOrdersUseCase _getOpenOrdersUseCase;
 
   Future<void> getOpenOrders(String address, String signature) async {
@@ -50,48 +46,6 @@ class ListOrdersCubit extends Cubit<ListOrdersState> {
           orderIdsLoading: [...previousState.orderIdsLoading],
         ),
       );
-    }
-  }
-
-  Future<bool> cancelOrder(
-    OrderEntity order,
-    String address,
-    String signature,
-  ) async {
-    final firstPreviousState = state;
-
-    if (firstPreviousState is ListOrdersLoaded &&
-        firstPreviousState.openOrders.contains(order)) {
-      emit(ListOrdersLoaded(
-        openOrders: [...firstPreviousState.openOrders],
-        orderIdsLoading: [
-          ...firstPreviousState.orderIdsLoading,
-          order.orderId,
-        ],
-      ));
-
-      final result = await _cancelOrderUseCase(
-        nonce: 0,
-        address: address,
-        orderId: order.orderId,
-        signature: signature,
-      );
-
-      final secondPreviousState = state;
-
-      if (secondPreviousState is ListOrdersLoaded) {
-        emit(ListOrdersLoaded(
-          openOrders: result.isRight()
-              ? ([...secondPreviousState.openOrders]..remove(order))
-              : [...secondPreviousState.openOrders],
-          orderIdsLoading: [...secondPreviousState.orderIdsLoading]
-            ..remove(order.orderId),
-        ));
-      }
-
-      return result.isRight();
-    } else {
-      return false;
     }
   }
 }
