+++ lib/features/landing/presentation/sub_views/trade_tab_view.dart
@@ -1,26 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:polkadex/common/configs/app_config.dart';
 import 'package:polkadex/common/cubits/account_cubit.dart';
 import 'package:polkadex/common/navigation/coordinator.dart';
 import 'package:polkadex/common/orderbook/presentation/cubit/orderbook_cubit.dart';
-import 'package:polkadex/common/widgets/polkadex_progress_error_widget.dart';
 import 'package:polkadex/features/coin/presentation/cubits/order_history_cubit.dart';
 import 'package:polkadex/features/landing/presentation/cubits/balance_cubit/balance_cubit.dart';
-import 'package:polkadex/features/landing/presentation/cubits/list_orders_cubit/list_orders_cubit.dart';
 import 'package:polkadex/features/landing/presentation/cubits/ticker_cubit/ticker_cubit.dart';
-import 'package:polkadex/features/landing/presentation/dialogs/trade_view_dialogs.dart';
 import 'package:polkadex/features/landing/presentation/providers/home_scroll_notif_provider.dart';
 import 'package:polkadex/features/landing/presentation/providers/trade_tab_provider.dart';
-import 'package:polkadex/features/landing/presentation/widgets/order_item_widget.dart';
 import 'package:polkadex/features/landing/presentation/cubits/place_order_cubit/place_order_cubit.dart';
 import 'package:polkadex/features/landing/presentation/widgets/trade_bottom_widget.dart';
 import 'package:polkadex/features/landing/presentation/widgets/place_order_widget.dart';
 import 'package:polkadex/features/landing/utils/token_utils.dart';
 import 'package:polkadex/common/orderbook/presentation/widgets/order_book_widget.dart';
 import 'package:polkadex/common/utils/colors.dart';
-import 'package:polkadex/common/utils/enums.dart';
 import 'package:polkadex/common/utils/extensions.dart';
 import 'package:polkadex/common/utils/styles.dart';
 import 'package:polkadex/common/widgets/build_methods.dart';
@@ -146,350 +140,6 @@ class _TradeTabViewState extends State<TradeTabView>
   void _scrollListener() {
     context.read<HomeScrollNotifProvider>().scrollOffset =
         _scrollController.offset;
-  }
-}
-
-/// The card includes buy sell tab controller
-class _ThisBuySellWidget extends StatefulWidget {
-  _ThisBuySellWidget({
-    required Key key,
-  }) : super(key: key);
-
-  @override
-  __ThisBuySellWidgetState createState() => __ThisBuySellWidgetState();
-}
-
-class __ThisBuySellWidgetState extends State<_ThisBuySellWidget>
-    with TickerProviderStateMixin {
-  late ValueNotifier<bool> _isOrdersExpanded;
-  late ValueNotifier<EnumBuySell> _buySellNotifier;
-  late ValueNotifier<EnumOrderTypes> _orderTypeSelNotifier;
-  late ValueNotifier<EnumTradeOrdersDisplayType?> _orderDisplayTypeNotifier;
-
-  @override
-  void initState() {
-    _orderTypeSelNotifier = ValueNotifier(EnumOrderTypes.market);
-    _isOrdersExpanded = ValueNotifier(false);
-    _orderDisplayTypeNotifier =
-        ValueNotifier<EnumTradeOrdersDisplayType?>(null);
-    _buySellNotifier = ValueNotifier(EnumBuySell.buy);
-    super.initState();
-    Future.microtask(() {
-      final tabController =
-          _ThisInheritedWidget.of(context)?.buySellTabController;
-      tabController?.addListener(() => _onSwapBuySellTab(context));
-    });
-  }
-
-  @override
-  void dispose() {
-    _buySellNotifier.dispose();
-    _isOrdersExpanded.dispose();
-    _orderTypeSelNotifier.dispose();
-    _orderDisplayTypeNotifier.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Consumer<TradeTabCoinProvider>(
-      builder: (context, coinProvider, child) =>
-          BlocBuilder<BalanceCubit, BalanceState>(
-        builder: (context, state) {
-          return Container(
-            decoration: BoxDecoration(
-              color: AppColors.color24252C,
-              borderRadius: BorderRadius.circular(20),
-              boxShadow: <BoxShadow>[
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.10),
-                  blurRadius: 30,
-                  offset: Offset(0.0, 20.0),
-                ),
-              ],
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                Padding(
-                  padding: const EdgeInsets.fromLTRB(24, 0, 30, 0),
-                  child: Row(
-                    children: [
-                      Expanded(
-                        child: ValueListenableBuilder<EnumBuySell>(
-                          valueListenable: _buySellNotifier,
-                          builder: (context, buyOrSell, child) {
-                            Color color;
-                            switch (buyOrSell) {
-                              case EnumBuySell.buy:
-                                color = AppColors.color0CA564;
-                                break;
-                              case EnumBuySell.sell:
-                                color = AppColors.colorE6007A;
-                                break;
-                            }
-                            return TabBar(
-                              isScrollable: true,
-                              labelStyle: tsS15W600CFF,
-                              unselectedLabelColor:
-                                  AppColors.colorFFFFFF.withOpacity(0.30),
-                              indicatorColor: color,
-                              indicatorSize: TabBarIndicatorSize.tab,
-                              tabs: <Tab>[
-                                Tab(
-                                  text:
-                                      'Buy ${TokenUtils.tokenIdToAcronym(coinProvider.tokenCoin.baseTokenId)}',
-                                ),
-                                Tab(
-                                  text:
-                                      'Sell ${TokenUtils.tokenIdToAcronym(coinProvider.tokenCoin.baseTokenId)}',
-                                ),
-                              ],
-                              controller: _ThisInheritedWidget.of(context)
-                                  ?.buySellTabController,
-                            );
-                          },
-                        ),
-                      ),
-                      InkWell(
-                        onTap: () => _onTapOrderType(context, coinProvider),
-                        child: Row(
-                          children: [
-                            Padding(
-                              padding: const EdgeInsets.only(right: 2.0),
-                              child: ValueListenableBuilder<EnumOrderTypes>(
-                                valueListenable: _orderTypeSelNotifier,
-                                builder: (context, type, _) => Text(
-                                  _getOrderTypeName(type),
-                                  style: tsS15W600CFF,
-                                ),
-                              ),
-                            ),
-                            Icon(
-                              Icons.keyboard_arrow_down,
-                              color: AppColors.colorFFFFFF,
-                              size: 16,
-                            ),
-                          ],
-                        ),
-                      )
-                    ],
-                  ),
-                ),
-                Padding(
-                  padding: const EdgeInsets.only(
-                    top: 4.0,
-                    left: 41,
-                    right: 41,
-                  ),
-                  child: InkWell(
-                    onTap: () {
-                      if (_orderDisplayTypeNotifier.value ==
-                          EnumTradeOrdersDisplayType.open) {
-                        if (_isOrdersExpanded.value) {
-                          _isOrdersExpanded.value = false;
-                        } else {
-                          _isOrdersExpanded.value = true;
-                        }
-                      } else {
-                        _orderDisplayTypeNotifier.value =
-                            EnumTradeOrdersDisplayType.open;
-                        if (!_isOrdersExpanded.value) {
-                          _isOrdersExpanded.value = true;
-                        }
-                      }
-                    },
-                    child: BlocBuilder<ListOrdersCubit, ListOrdersState>(
-                      builder: (context, state) {
-                        return Align(
-                          alignment: Alignment.center,
-                          child: state is ListOrdersError
-                              ? PolkadexErrorRefreshWidget(
-                                  textSize: 16,
-                                  onRefresh: () => context
-                                      .read<ListOrdersCubit>()
-                                      .getOpenOrders(
-                                        context
-                                            .read<AccountCubit>()
-                                            .accountAddress,
-                                        context
-                                            .read<AccountCubit>()
-                                            .accountSignature,
-                                      ),
-                                )
-                              : Stack(
-                                  alignment: Alignment.topRight,
-                                  children: [
-                                    Padding(
-                                      padding: const EdgeInsets.only(
-                                          top: 8, right: 12),
-                                      child: Text(
-                                        "Open Orders",
-                                        style: state is ListOrdersLoaded &&
-                                                state.openOrders.isEmpty
-                                            ? tsS15W600CABB2BC
-                                            : tsS15W600CFF,
-                                      ),
-                                    ),
-                                    if (state is! ListOrdersError)
-                                      Container(
-                                        decoration: BoxDecoration(
-                                          color: AppColors.colorE6007A,
-                                          shape: BoxShape.circle,
-                                        ),
-                                        padding: EdgeInsets.all(
-                                            state is ListOrdersLoading
-                                                ? 2.5
-                                                : 4),
-                                        margin:
-                                            const EdgeInsets.only(bottom: 4),
-                                        child: state is ListOrdersLoaded
-                                            ? Text('${state.openOrders.length}',
-                                                style: tsS10W500CFF)
-                                            : SizedBox(
-                                                width: 6.0,
-                                                height: 6.0,
-                                                child:
-                                                    CircularProgressIndicator(
-                                                  strokeWidth: 1.5,
-                                                  color: Colors.white,
-                                                ),
-                                              ),
-                                      ),
-                                  ],
-                                ),
-                        );
-                      },
-                    ),
-                  ),
-                ),
-                ValueListenableBuilder<bool>(
-                  valueListenable: _isOrdersExpanded,
-                  builder: (context, isShow, child) => AnimatedSize(
-                    duration: AppConfigs.animDurationSmall,
-                    alignment: Alignment.topCenter,
-                    child: ValueListenableBuilder<EnumTradeOrdersDisplayType?>(
-                      valueListenable: _orderDisplayTypeNotifier,
-                      builder: (context, orderDisplayType, child) {
-                        if (!(isShow)) return SizedBox(height: 15);
-                        return _ThisOpenOrderExpandedWidget(
-                          type: orderDisplayType,
-                        );
-                      },
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          );
-        },
-      ),
-    );
-  }
-
-  /// The callback listener for otder type selection
-  void _onTapOrderType(
-    BuildContext context,
-    TradeTabCoinProvider coinProvider,
-  ) {
-    showOrderTypeDialog(
-      context: context,
-      selectedIndex: _orderTypeSelNotifier.value,
-      onItemSelected: (index) {
-        _orderTypeSelNotifier.value = index;
-
-        if (index == EnumOrderTypes.market) {
-          context.read<TickerCubit>().getLastTicker(
-                leftTokenId: coinProvider.tokenCoin.baseTokenId,
-                rightTokenId: coinProvider.tokenCoin.pairTokenId,
-              );
-        }
-      },
-    );
-  }
-
-  /// Returns the order type string from the [type]
-  String _getOrderTypeName(EnumOrderTypes type) {
-    switch (type) {
-      case EnumOrderTypes.market:
-        return "Market Order";
-      case EnumOrderTypes.limit:
-        return "Limit Order";
-      case EnumOrderTypes.stop:
-        return "Stop Order";
-    }
-  }
-
-  void _onSwapBuySellTab(BuildContext context) {
-    final tabController =
-        _ThisInheritedWidget.of(context)?.buySellTabController;
-
-    _buySellNotifier.value = EnumBuySell.values[tabController!.index];
-  }
-}
-
-/// This widget will be shown when the user click on open orders
-class _ThisOpenOrderExpandedWidget extends StatelessWidget {
-  final EnumTradeOrdersDisplayType? type;
-  const _ThisOpenOrderExpandedWidget({
-    required this.type,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        SizedBox(height: 8),
-        Align(
-          alignment: Alignment.center,
-          child: Container(
-            width: 104,
-            height: 4,
-            margin: const EdgeInsets.only(bottom: 2.5, left: 40, right: 40),
-            decoration: BoxDecoration(
-              color: AppColors.colorE6007A,
-              borderRadius: BorderRadius.circular(10),
-            ),
-          ),
-        ),
-        SizedBox(height: 8),
-        BlocBuilder<ListOrdersCubit, ListOrdersState>(
-          builder: (context, state) {
-            if (state is ListOrdersLoaded) {
-              return Column(
-                children: state.openOrders
-                    .map<Widget>(
-                      (order) => OrderItemWidget(
-                        order: order,
-                        isProcessing:
-                            state.orderIdsLoading.contains(order.orderId),
-                        onTapClose: () async {
-                          final cancelSuccess = await context
-                              .read<ListOrdersCubit>()
-                              .cancelOrder(
-                                order,
-                                context.read<AccountCubit>().accountAddress,
-                                context.read<AccountCubit>().accountSignature,
-                              );
-
-                          buildAppToast(
-                              msg: cancelSuccess
-                                  ? 'Order cancelled successfully'
-                                  : 'Order cancel failed. Please try again',
-                              context: context);
-                        },
-                      ),
-                    )
-                    .toList(),
-              );
-            }
-
-            return Container();
-          },
-        ),
-      ],
-    );
   }
 }
