+++ test/features/landing/presentation/order_history_cubit_test.dart
@@ -5,6 +5,7 @@ import 'package:polkadex/common/network/error.dart';
 import 'package:polkadex/common/orders/data/models/fee_model.dart';
 import 'package:polkadex/common/orders/data/models/order_model.dart';
 import 'package:polkadex/common/orders/domain/entities/fee_entity.dart';
+import 'package:polkadex/common/orders/domain/usecases/cancel_order_usecase.dart';
 import 'package:polkadex/common/orders/domain/usecases/get_orders_usecase.dart';
 import 'package:polkadex/common/utils/enums.dart';
 import 'package:polkadex/features/coin/presentation/cubits/order_history_cubit.dart';
@@ -12,8 +13,11 @@ import 'package:test/test.dart';
 
 class _MockGetOrdersUsecase extends Mock implements GetOrdersUseCase {}
 
+class _MockCancelOrderUsecase extends Mock implements CancelOrderUseCase {}
+
 void main() {
-  late _MockGetOrdersUsecase _mockOrdersUsecase;
+  late _MockGetOrdersUsecase _mockGetOrdersUsecase;
+  late _MockCancelOrderUsecase _mockCancelOrderUsecase;
   late OrderHistoryCubit cubit;
   late String orderId;
   late String mainAcc;
@@ -34,10 +38,12 @@ void main() {
   late OrderModel order;
 
   setUp(() {
-    _mockOrdersUsecase = _MockGetOrdersUsecase();
+    _mockGetOrdersUsecase = _MockGetOrdersUsecase();
+    _mockCancelOrderUsecase = _MockCancelOrderUsecase();
 
     cubit = OrderHistoryCubit(
-      getOrdersUseCase: _mockOrdersUsecase,
+      getOrdersUseCase: _mockGetOrdersUsecase,
+      cancelOrderUseCase: _mockCancelOrderUsecase,
     );
 
     orderId = '786653432';
@@ -84,7 +90,7 @@ void main() {
         'Orders fetched successfully',
         build: () {
           when(
-            () => _mockOrdersUsecase(
+            () => _mockGetOrdersUsecase(
               address: any(named: 'address'),
               signature: any(named: 'signature'),
             ),
@@ -103,7 +109,10 @@ void main() {
         },
         expect: () => [
           OrderHistoryLoading(),
-          OrderHistoryLoaded(orders: [order]),
+          OrderHistoryLoaded(
+            orders: [order],
+            orderIdsLoading: [],
+          ),
         ],
       );
 
@@ -111,7 +120,7 @@ void main() {
         'Orders fetch fail',
         build: () {
           when(
-            () => _mockOrdersUsecase(
+            () => _mockGetOrdersUsecase(
               address: any(named: 'address'),
               signature: any(named: 'signature'),
             ),
